requirements.txt
Django==5.1.1
gunicorn==21.2.0
whitenoise==6.6.0
djangorestframework==3.15.1
python-dotenv==1.0.1
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.3.1
drf-spectacular-sidecar==2024.12.1
django-cors-headers==4.6.0
drf-yasg==2024.2